requirements.txt
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
